pyproject.toml
[project]
name = "django-template"
version = "0.1.0"
description = ""
authors = [{ name = "harold", email = "harold.k.basa@gmail.com" }]
requires-python = ">=3.13"
readme = "README.md"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "django-allauth[socialaccount]>=65.5.0",
    "django-ninja>=1.3.0",
    "django-ninja-extra>=0.22.6",
    "psycopg[binary] (>=3.2.6,<4.0.0)",
    "pydantic[email]>=2.10.6",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.1.3",
    "mypy>=1.15.0",
    "ruff>=0.11.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.uv]
package = false

[tool.ruff]
line-length = 100
src = ["app"]

[tool.ruff.lint]
# rules can be found here: https://docs.astral.sh/ruff/rules/
select = [
    "F",    # pyflakes
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "I",    # isort
    "ERA",  # eradicate
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "T10",  # flake8-django
    "T20",  # flake8-print
    "B",    # flake8-bugbear
    "G",    # flake8-logging-format
    "RUF"   # Ruff-specific rules
]

# allow `isort` fixes when `--fix` is provided
fixable = ["I"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "E402"     # import violations
]
"**/migrations/*.py" = [
    "RUF012",   # mutable-class-default
    "E501"      # line-too-long  (remove when ruff supports formatting long strings)
]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = true
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.mypy]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
